🐛 fix: fix bug

+++ events/messageCreate.js
@@ -15,7 +15,7 @@ module.exports = {
     if (message.channel.id === "912033913843630080" && !message.author.bot) {
       require("../modules/vip").create(message);
     } else if (
-      message.author.bot &&
+      message.channel.id === "912033913843630080" && message.author.bot &&
       message.author.id !== message.client.user.id
     )
       message.delete();
@@ -27,7 +27,8 @@ module.exports = {
     ) {
       require("../modules/vip").syori(message);
     } else if (
-      message.author.bot &&
+      message.channel instanceof Discord.ThreadChannel &&
+      message.channel.parent.id === "912033913843630080" && message.author.bot &&
       message.author.id !== message.client.user.id
     )
       message.delete();
